+++ rs-commerce/src/constants/constants.ts
@@ -69,6 +69,7 @@ export const ROUTE_PATH = {
 export const LOCAL_STORAGE_TOKEN = 'hurricane_access_token';
 
 export const LOCAL_STORAGE_ANONYM_CART_ID = 'hurricane_anonym_cart';
+export const LOCAL_STORAGE_AUTH_CART_ID = 'hurricane_auth_cart';
 
 export const STATUS = {
   CODE_400: 400,
