+++ rs-commerce/src/views/productDetails/ProductDetails.tsx
@@ -1,15 +1,20 @@
 import { useParams } from 'react-router-dom';
 import { useEffect, useState } from 'react';
-import { apiCreateCart, apiGetCart, apiGetOneCategory, apiGetProductDetails } from 'api/api';
+import { /* apiCreateCart, apiGetCart, */ apiGetOneCategory, apiGetProductDetails } from 'api/api';
 import Loader from 'components/loader/loader';
 import getRequestErrorMessage from 'utils/utils';
 import { useAppDispatch, useAppSelector } from 'hooks/typed-react-redux-hooks';
 import { productDetailSliceActions } from 'redux/slices/product-detail-slice';
-import { apiAuthSelector, cartSelector, productDetailsSelector } from 'redux/selectors';
+import { /* apiAuthSelector, cartSelector, */ productDetailsSelector } from 'redux/selectors';
 import { ErrorResponse } from '@commercetools/importapi-sdk';
 import toast from 'react-hot-toast';
-import { LOCAL_STORAGE_ANONYM_CART_ID, ROUTE_PATH } from 'constants/constants';
-import { cartActions } from 'redux/slices/cart-slice';
+import {
+  // LOCAL_STORAGE_ANONYM_CART_ID,
+  // LOCAL_STORAGE_AUTH_CART_ID,
+  // LOCAL_STORAGE_TOKEN,
+  ROUTE_PATH,
+} from 'constants/constants';
+// import { cartActions } from 'redux/slices/cart-slice';
 import { Breadcrumbs, DescriptionBlock, ImageBlock } from './components';
 import style from './style.module.css';
 
@@ -17,8 +22,8 @@ function ProductDetails() {
   const { id } = useParams();
   const dispatch = useAppDispatch();
   const { productDetail } = useAppSelector(productDetailsSelector);
-  const { cartData } = useAppSelector(cartSelector);
-  const { isAuth } = useAppSelector(apiAuthSelector);
+  // const { cartData } = useAppSelector(cartSelector);
+  // const { isAuth } = useAppSelector(apiAuthSelector);
   const [isLoading, setIsLoading] = useState(true);
   const [errorMessage, setErrorMessage] = useState('');
   const [categoryLink, setCategoryLink] = useState<{ name: string; path: string }[]>([]);
@@ -86,31 +91,36 @@ function ProductDetails() {
     }
   }, [productDetail]);
 
-  const getCartData = async () => {
-    try {
-      let data = null;
-      let cartId = localStorage.getItem(LOCAL_STORAGE_ANONYM_CART_ID);
+  // const getCartData = async () => {
+  //   try {
+  //     let data = null;
+  //     let cartId = isAuth
+  //       ? localStorage.getItem(LOCAL_STORAGE_AUTH_CART_ID)
+  //       : localStorage.getItem(LOCAL_STORAGE_ANONYM_CART_ID);
 
-      if (!isAuth && !cartId) {
-        const newCart = await apiCreateCart();
-        cartId = newCart.body.id;
-        localStorage.setItem(LOCAL_STORAGE_ANONYM_CART_ID, cartId);
-      }
+  //     if (!cartId) {
+  //       const newCart = await apiCreateCart();
+  //       cartId = newCart.body.id;
+  //       // after deletion of anonymous cart in cart page remove it from local storage
+  //       localStorage.setItem(
+  //         isAuth ? LOCAL_STORAGE_AUTH_CART_ID : LOCAL_STORAGE_ANONYM_CART_ID,
+  //         cartId,
+  //       );
+  //     }
+  //     data = await apiGetCart(cartId);
 
-      data = await apiGetCart(cartId);
+  //     if (data?.body) dispatch(cartActions.setCartData(data?.body));
+  //   } catch (e) {
+  //     const error = getRequestErrorMessage(e.code);
+  //     toast.error(error);
+  //   }
+  // };
 
-      if (data?.body) dispatch(cartActions.setCartData(data?.body));
-    } catch (e) {
-      const error = getRequestErrorMessage(e.code);
-      toast.error(error);
-    }
-  };
-
-  useEffect(() => {
-    if (!cartData) {
-      getCartData();
-    }
-  }, [cartData]);
+  // useEffect(() => {
+  //   if (!cartData) {
+  //     getCartData();
+  //   }
+  // }, [cartData]);
 
   if (isLoading) return <Loader isShow />;
   if (errorMessage) return <div className={`container ${style['error']}`}>{errorMessage}</div>;
